requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
decorator==5.1.1
dj-database-url==1.3.0
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
forex-python==1.8
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.1.0
incremental==22.10.0
jsonpickle==3.0.1
Markdown==3.4.2
msgpack==1.0.5
Pillow==9.4.0
psycopg==3.1.8
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
pytz==2022.7.1
redis==4.5.4
requests==2.28.2
service-identity==21.1.0
simplejson==3.18.4
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
urllib3==1.26.14
validators==0.20.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
